+++ share/noba-web/frontend/src/main.js
@@ -3,6 +3,12 @@ import { createPinia } from 'pinia'
 import router from './router'
 import App from './App.vue'
 import '@fortawesome/fontawesome-free/css/all.min.css'
+import '@fontsource/chakra-petch/400.css'
+import '@fontsource/chakra-petch/500.css'
+import '@fontsource/chakra-petch/600.css'
+import '@fontsource/chakra-petch/700.css'
+import '@fontsource/jetbrains-mono/400.css'
+import '@fontsource/jetbrains-mono/500.css'
 import './assets/styles/global.css'
 
 const app = createApp(App)